requirements.txt
pandas>=2.0.0
aiohttp>=3.8.4
requests>=2.28.2
ipykernel>=6.22.0
scikit-learn>=1.2.2
joblib>=1.2.0
websocket-client>=1.5.1
httpx